@@ -2,33 +2,36 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": ["framework", "laravel"],
+    "keywords": [
+        "framework",
+        "laravel"
+    ],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^8.0",
         "doctrine/dbal": "^3.6",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
+        "guzzlehttp/guzzle": "^7.2",
         "jeroennoten/laravel-adminlte": "^3.7",
-        "laravel/framework": "^8.65",
-        "laravel/sanctum": "^2.11",
+        "laravel/framework": "^9.52",
+        "laravel/sanctum": "^3.0",
         "laravel/telescope": "^4.9",
-        "laravel/tinker": "^2.5",
+        "laravel/tinker": "^2.7",
         "laravel/ui": "^3.4",
         "maatwebsite/excel": "^3.1",
         "spatie/db-dumper": "^2.21",
-        "spatie/laravel-backup": "^6.16",
-        "yajra/laravel-datatables-oracle": "^9.18",
-        "laravel/helpers": "^1.3"
+        "spatie/laravel-backup": "^8.0",
+        "yajra/laravel-datatables-oracle": "^9.19",
+        "laravel/helpers": "^1.5",
+        "league/flysystem-aws-s3-v3": "^3.0"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.6",
-        "facade/ignition": "^2.5",
         "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.0.1",
+        "laravel/sail": "^1.15",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^5.10",
-        "phpunit/phpunit": "^9.5.10"
+        "nunomaduro/collision": "^6.3",
+        "phpunit/phpunit": "^9.5.10",
+        "spatie/laravel-ignition": "^1.4"
     },
     "autoload": {
         "psr-4": {
