@@ -14,8 +14,8 @@ class UserSeeder extends Seeder
      */
     public function run()
     {
-        $users = array(
-            array(
+        $users = [
+            [
                 "id_grup" => 1,
                 "name" => "Eddie Ridwan",
                 "username" => "eddieridwan",
@@ -24,8 +24,8 @@ class UserSeeder extends Seeder
                 "password" => '$2y$10$xyNmjtuWL3.apmIgQGZ2y.c8X908ym8PlbkZQPCB4iJoVHq90Fv8q',
                 "token" => null,
                 "created_at" => now(),
-            ),
-            array(
+            ],
+            [
                 "id_grup" => 1,
                 "name" => "Herry Wanda",
                 "username" => "herrywanda",
@@ -34,8 +34,8 @@ class UserSeeder extends Seeder
                 "password" => '$2y$10$b6IFYRt9th6nr1S2dfbb4epFIFpy.QcXbN.iB7SzMJuFO4BjePdq6',
                 "token" => null,
                 "created_at" => now(),
-            ),
-            array(
+            ],
+            [
                 "id_grup" => 1,
                 "name" => "Rudi Purwanto",
                 "username" => "roaddee",
@@ -44,8 +44,8 @@ class UserSeeder extends Seeder
                 "password" => '$2y$10$V7iREjPiUnjvVkIOp9iPIeV72wb/k0z4NBXwT9raUT4XpvVYnWvOa',
                 "token" => 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6bnVsbCwidGltZXN0YW1wIjoxNjE0MjM5ODQzfQ.HlzpyJgG431dw17idGuU70b1FJXW7ZrmRZRzsC9jyIU',
                 "created_at" => now(),
-            ),
-            array(
+            ],
+            [
                 "id_grup" => 1,
                 "name" => "Muhammad Ihsan",
                 "username" => "muhammadihsan",
@@ -54,8 +54,8 @@ class UserSeeder extends Seeder
                 "password" => "$2y$10$1Nikfgo/0yYjQGVDb1vGhegoXyFTiqYfCaTx4Xc8RlB4zcK6kSPie",
                 "token" => null,
                 "created_at" => now(),
-            ),
-            array(
+            ],
+            [
                 "id_grup" => 1,
                 "name" => "Muhammad AI",
                 "username" => "aiskematik",
@@ -64,8 +64,8 @@ class UserSeeder extends Seeder
                 "password" => "$2y$10$/TyNRWV5NJN4zekzIaEMbO/mzwc9HunTmNoUAnaKl9x1AMbd3/4vW",
                 "token" => null,
                 "created_at" => now(),
-            ),
-            array(
+            ],
+            [
                 "id_grup" => 1,
                 "name" => "Agung Sugiarto",
                 "username" => "agungsugiarto",
@@ -74,8 +74,8 @@ class UserSeeder extends Seeder
                 "password" => '$2y$10$u9P4w37ps8jS1mwFdgsWuOYdRha2UYTRyhLaKCUpqj3I4dl9PY4IW',
                 "token" => null,
                 "created_at" => now(),
-            ),
-            array(
+            ],
+            [
                 "id_grup" => 1,
                 "name" => "Andi Fahruddin Akas",
                 "username" => "andifahruddinakas",
@@ -84,8 +84,8 @@ class UserSeeder extends Seeder
                 "password" => '$2y$10$frHX2pWN3XQiZeljnpDtnOkMqWpTnajYVjFeKNR9K04oxlkyqpf9u',
                 "token" => null,
                 "created_at" => now(),
-            ),
-        );
+            ],
+        ];
 
         DB::table('users')->insert($users);
     }
