@@ -130,7 +130,7 @@ if (! function_exists('lastrelease')) {
             ->throw();
 
             return json_decode($response->body());
-        } catch (\Throwable $th) {
+        } catch (Throwable $th) {
             return false;
         }
     }
