@@ -5,7 +5,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class Addtablepengaturan extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
@@ -49,4 +49,4 @@ class Addtablepengaturan extends Migration
     {
         Schema::dropIfExists('pengaturan_aplikasi');
     }
-}
+};
